+++ src/HsmEventDispatcherGLibmm.cpp
@@ -2,24 +2,32 @@
 // Distributed under MIT license. See file LICENSE for detail
 
 #include "HsmEventDispatcherGLibmm.hpp"
+#include "logging.hpp"
+
+#undef __TRACE_CLASS__
+#define __TRACE_CLASS__                         "HsmEventDispatcherGLibmm"
 
 HsmEventDispatcherGLibmm::HsmEventDispatcherGLibmm()
-    : mDispatcher(std::make_shared<Glib::Dispatcher>())
+    : mDispatcher(std::make_unique<Glib::Dispatcher>())
 {
+    __TRACE_CALL_DEBUG__();
 }
 
 HsmEventDispatcherGLibmm::HsmEventDispatcherGLibmm(const Glib::RefPtr<Glib::MainContext>& context)
-    : mDispatcher(std::make_shared<Glib::Dispatcher>(context))
+    : mDispatcher(std::make_unique<Glib::Dispatcher>(context))
 {
+    __TRACE_CALL_DEBUG__();
 }
 
 HsmEventDispatcherGLibmm::~HsmEventDispatcherGLibmm()
 {
+    __TRACE_CALL_DEBUG__();
     unregisterAllEventHandlers();
 }
 
 int HsmEventDispatcherGLibmm::registerEventHandler(std::function<void(void)> handler)
 {
+    __TRACE_CALL_DEBUG__();
     int id = INVALID_HSM_DISPATCHER_HANDLER_ID;
 
     if (mDispatcher)
@@ -33,6 +41,7 @@ int HsmEventDispatcherGLibmm::registerEventHandler(std::function<void(void)> han
 
 void HsmEventDispatcherGLibmm::unregisterEventHandler(const int handlerId)
 {
+    __TRACE_CALL_DEBUG_ARGS__("handlerId=%d", handlerId);
     auto it = mEventHandlers.find(handlerId);
 
     if (it != mEventHandlers.end())
@@ -44,6 +53,7 @@ void HsmEventDispatcherGLibmm::unregisterEventHandler(const int handlerId)
 
 void HsmEventDispatcherGLibmm::emit()
 {
+    __TRACE_CALL_DEBUG__();
     if (mDispatcher)
     {
         mDispatcher->emit();