+++ tests/mainSTD.cpp
@@ -1,6 +1,7 @@
 #include <gmock/gmock.h>
 #include <gtest/gtest.h>
 #include "logging.hpp"
+#include "TestsCommon.hpp"
 
 __TRACE_PREINIT__();
 
@@ -9,6 +10,7 @@ int main(int argc, char** argv)
     __TRACE_INIT__();
 
     ::testing::InitGoogleMock(&argc, argv);
+    configureGTest();
 
     return RUN_ALL_TESTS();
 }